feat(Marketing Pages): refactor service marketing landing page structure and enhance animations
- Removed the legacy MarketingLandingPage component to streamline the codebase.
- Introduced a modular structure for the ServiceMarketingLandingPage, breaking down sections into reusable components (Hero, Benefits, Features, Process, Testimonials, Deliverables, Contact Form, CTA).
- Implemented a custom hook for managing animations and scroll effects, improving performance and maintainability.
- Enhanced user experience with improved animations and transitions across various sections.
- Added new data structures for service-specific content, including features and testimonials, to support dynamic rendering.
