+++ app/(site)/marketing/components/sections/FeaturesSection.tsx
@@ -1,0 +1,67 @@
+"use client";
+
+import { motion } from "framer-motion";
+import { CheckCircle } from "lucide-react";
+import { Card, CardContent } from "@/components/ui/card";
+
+interface FeaturesSectionProps {
+  service: {
+    name: string;
+    features: string[];
+  };
+  featuresRef: React.RefObject<HTMLDivElement>;
+  featuresInView: boolean;
+}
+
+export function FeaturesSection({
+  service,
+  featuresRef,
+  featuresInView,
+}: FeaturesSectionProps) {
+  return (
+    <motion.section ref={featuresRef} className="py-20 bg-white">
+      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+        <motion.div
+          initial={{ opacity: 0, y: 50 }}
+          animate={featuresInView ? { opacity: 1, y: 0 } : {}}
+          transition={{ duration: 0.8 }}
+          className="text-center mb-16"
+        >
+          <h2 className="text-4xl md:text-5xl font-bold text-gray-900 mb-6">
+            What&apos;s Included in Your {service.name} Package
+          </h2>
+          <p className="text-xl text-gray-600">
+            Everything you need to succeed with {service.name.toLowerCase()}
+          </p>
+        </motion.div>
+
+        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
+          {service.features.map((feature, index) => (
+            <motion.div
+              key={index}
+              initial={{ opacity: 0, y: 50 }}
+              whileInView={{ opacity: 1, y: 0 }}
+              transition={{ delay: index * 0.1, duration: 0.8 }}
+              viewport={{ once: true }}
+            >
+              <Card className="p-6 h-full border-l-4 border-l-coral-orange">
+                <CardContent className="p-0">
+                  <div className="flex items-start space-x-4">
+                    <div className="flex-shrink-0">
+                      <CheckCircle className="w-6 h-6 text-coral-orange" />
+                    </div>
+                    <div>
+                      <h3 className="text-lg font-semibold text-gray-900 mb-2">
+                        {feature}
+                      </h3>
+                    </div>
+                  </div>
+                </CardContent>
+              </Card>
+            </motion.div>
+          ))}
+        </div>
+      </div>
+    </motion.section>
+  );
+}
